requirements.txt
alembic==1.15.2
bcrypt==4.3.0
certifi==2025.4.26
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
Mako==1.3.9
MarkupSafe==3.0.2
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.1.0
sentry-sdk==2.28.0
SQLAlchemy==2.0.40
typing_extensions==4.13.0
urllib3==2.4.0
